Change datastructure for regions, add tests
Accessing the regions as a vector with the rank as index is cumbersome
for many computations that rely on the index-order (i.e. ordering of
globalStartIndex). Using a permutation to transform ranks into
index-order was not fruitful either.

The new approach uses two `std::maps`: the `region` map that maps from
rank to region and `start_indices` which maps from start index to rank.
Most importantly, both maps only contain ranks that actually have
elements assigned, making dealing with zero-assignments much easier
since they do not need to be skipped in all computations.